UTF8 Bug fixed

+++ lib/Helper.class.php
@@ -215,6 +215,9 @@ class Helper {
         $db = self::getDbObject();
         $db->query("CREATE DATABASE `{$config['db']}`  DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci;");
         $tmpdb = self::getDbObject($config);
+        if (!$tmpdb->set_charset("utf8")) {
+            throw new \Exception(sprintf("Ошибка установки CHARACTER SET utf8: %s\n", $tmpdb->error));
+        }
         register_shutdown_function(function() use($config, $tmpdb) {
                     $tmpdb->query("DROP DATABASE `{$config['db']}`");
                     Output::verbose("Временная база данных {$config['db']} была удалена");
